cheese-48 MemberService Interface 작성
